Move error to crate root.
