+++ deepwell/src/cache.rs
@@ -18,7 +18,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-use crate::services::Result;
+use crate::error::Result;
 use rsmq_async::Rsmq;
 
 /// Creates primary `redis::Client` instance.
