+++ deepwell/src/endpoints/mod.rs
@@ -30,13 +30,14 @@
 #[allow(unused_imports)]
 mod prelude {
     pub use crate::api::ServerState;
+    pub use crate::error::{Error as ServiceError, Result, StdResult};
     pub use crate::services::{
         AliasService, BlobService, CaddyService, CategoryService, DomainService,
-        Error as ServiceError, FileRevisionService, FileService, LinkService,
-        MessageReportService, MessageService, MfaService, PageRevisionService,
-        PageService, ParentService, RelationService, RenderService, Result, ScoreService,
-        ServiceContext, SessionService, SettingsService, SiteService, StdResult,
-        TextBlockService, TextService, UserService, ViewService, VoteService,
+        FileRevisionService, FileService, LinkService, MessageReportService,
+        MessageService, MfaService, PageRevisionService, PageService, ParentService,
+        RelationService, RenderService, ScoreService, ServiceContext, SessionService,
+        SettingsService, SiteService, TextBlockService, TextService, UserService,
+        ViewService, VoteService,
     };
     pub use jsonrpsee::types::params::Params;
     pub use std::convert::TryFrom;
