+++ deepwell/src/error/mod.rs
@@ -1,5 +1,5 @@
 /*
- * services/error/mod.rs
+ * error/mod.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2026 Wikijump Team
