+++ deepwell/src/error/new.rs
@@ -1,5 +1,5 @@
 /*
- * services/error/contextual.rs
+ * error/new.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2026 Wikijump Team
