+++ deepwell/src/license.rs
@@ -20,8 +20,8 @@
 
 //! Constant data for licenses usable by Wikijump sites.
 
+use crate::error::Result;
 use crate::locales::Localizations;
-use crate::services::Result;
 use fluent::FluentArgs;
 use unic_langid::LanguageIdentifier;
