+++ deepwell/src/services/site/service.rs
@@ -20,6 +20,7 @@
 
 use super::prelude::*;
 use crate::constants::SYSTEM_USER_ID;
+use crate::error::Error;
 use crate::models::sea_orm_active_enums::{AliasType, UserType};
 use crate::models::site::{self, Entity as Site, Model as SiteModel};
 use crate::services::alias::CreateAlias;
@@ -27,7 +28,7 @@ use crate::services::audit::{AuditEvent, AuditService, SiteFields};
 use crate::services::domain::{DEFAULT_SITE_SLUG, DomainService};
 use crate::services::relation::CreateSiteUser;
 use crate::services::user::{CreateUser, UpdateUserBody};
-use crate::services::{AliasService, Error, RelationService, UserService};
+use crate::services::{AliasService, RelationService, UserService};
 use crate::utils::validate_locale;
 use ftml::layout::Layout;
 use ref_map::*;
